requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
patchright==1.50.0
tenacity==8.2.3
python-dotenv==1.0.1
openai==2.32.0
pytest==8.3.3
